+++ webmarket/pom.xml
@@ -176,8 +176,8 @@
 					<argLine>${surefireArgLine}</argLine>
 					<!-- Skips unit tests if the value of skip.unit.tests property is true -->
 					<skipTests>${skip.unit.tests}</skipTests>
-					<!-- Excludes integration tests when unit tests are run. -->
-					<parallelOptimized>false</parallelOptimized>
+					<!-- Excludes integration tests when unit tests are run.
+					<parallelOptimized>false</parallelOptimized>-->
 					<excludes>
 						<exclude>**/IT*.java</exclude>
 					</excludes>
